backend/requirements.txt
django
djangorestframework
djoser
djangorestframework-simplejwt
psycopg2-binary
gunicorn
django-cors-headers